+++ cuda/tomo_cuda.hpp
@@ -1,5 +1,0 @@
-#include "forward_project.hpp"
-#include "sart.hpp"
-#include "sirt.hpp"
-#include "bridges.hpp"
-#include "device_data.hpp"