+++ python/tomos/__init__.py
@@ -1,0 +1,8 @@
+from . import py_tomos
+from tomos.py_tomos import *
+
+from . import util
+from tomos.util import *
+
+__version__ = '0.0.1'
+