+++ src/phantoms.cpp
@@ -1,7 +1,0 @@
-#include "tomos/tomos.hpp"
-
-namespace tomo {
-
-int f() { return 0; }
-
-} // namespace tomo