+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.2.0
+
+2017-10-01
+
+- Add tools for partitioning.
+- Improve build system.
+- Remove obsolete and unclear examples.
+- Remove CUDA code (GPU support may return in a future release in a revised form)
+
 ## 0.1.0
 
 2017-06-01